Изменения в ProjectController.cs, новые запросы sql, новые контексты

+++ src/server/Organization/Controllers/ProjectController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System.Diagnostics.CodeAnalysis;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Organization.Controllers.Common;
 using Organization.Models.Contexts;
@@ -11,6 +12,7 @@ namespace Organization.Controllers;
 //TODO Замени заглушки на ответы от DB
 [ApiController]
 [Route("Organizations/{OrganizationId}")]
+[SuppressMessage("ReSharper", "InconsistentNaming")]
 public class ProjectController : ControllerBase
 {
     private readonly ILogger<ProjectController> _logger;
@@ -30,7 +32,12 @@ public class ProjectController : ControllerBase
     [HttpGet("Projects")]
     public IActionResult GetProjects(ulong OrganizationId)
     {
-        var entities = _projectContext.Projects.Where(x => x.organization_id == OrganizationId);
+        var entities = _projectContext.Projects
+            .Where(x => x.organization_id == OrganizationId)
+            .Include(x => x.boards)
+            .ThenInclude(x => x.Columns)
+            .ThenInclude(x => x.Tasks)
+            .ThenInclude(x => x.Components);
         return ResponseGenerator.Ok(value: entities.ToList());
     }
 
@@ -47,6 +54,8 @@ public class ProjectController : ControllerBase
     [HttpPut("Projects")]
     public IActionResult PutProjects(ulong OrganizationId, Project project)
     {
+        if (project.organization_id != OrganizationId)
+            return ResponseGenerator.NotFound();
         _projectContext.Projects.Add(project);
         _projectContext.SaveChanges();
         return ResponseGenerator.Ok(value: _projectContext.Projects.ToList());
@@ -61,7 +70,11 @@ public class ProjectController : ControllerBase
     {
         //TODO 
         // var context = new BoardContext();
-        return ResponseGenerator.Ok(value: _boardContext.Boards.ToList());
+        return ResponseGenerator.Ok(value: _boardContext.Boards.
+            Include(x=> x.Columns).
+            ThenInclude(x=> x.Tasks).
+            ThenInclude(x =>x.Components).
+            ToList());
         
         throw new NotImplementedException();
     }
@@ -81,7 +94,11 @@ public class ProjectController : ControllerBase
     {
         //TODO 
         var table = _boardContext.Boards;
-        var entities = table.Where(x => x.id == BoardId);
+        if (!_projectContext.Projects.Any(x => x.organization_id == OrganizationId && x.id == ProjectId))
+            return ResponseGenerator.NotFound();
+        var entities = table.Where(x => x.id == BoardId && x.ProjectId == ProjectId);
+        if (!entities.Any())
+            return ResponseGenerator.NotFound();
         table.RemoveRange(entities);
         _boardContext.SaveChanges();
         return ResponseGenerator.Ok(value: table.ToList());
